refactor(pagination): resolve svelte imports
